requirements.txt
numba==0.56.4
shap==0.42.0
pyoperon==0.3.4
redis==3.5.3
scikit-learn==1.2.2
celery
graphviz==0.20.1
pandas==2.0.3
pyarrow==12.0.1
matplotlib==3.3.4
pytest==7.4.0
sseclient==0.0.27
gplearn==0.4.2
lime==0.2.0.1
git+https://github.com/leauferreira/GpX@holow_gpx
alibi==0.9.4
torch==2.0.1
torchvision==0.15.2
captum==0.6.0
confluent-kafka
kombu
python-dotenv
SQLAlchemy
psycopg2-binary